# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E3Di & Jaan LED Screen IoT Ad-Network Platform API",
    "description": "Comprehensive REST & IoT API documentation for the E3Di Smart Pole, Digital Billboard Network, Edge Gateways, CCTV surveillance, and Emergency SOS infrastructure.",
    "version": "2.0.0",
    "contact": {
      "name": "E3Di Platform Support",
      "url": "https://e3di.org"
    }
  },
  "servers": [
    {
      "url": "https://api.e3di.org",
      "description": "Production LED CPU API Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server (Port 5000)"
    },
    {
      "url": "http://localhost:5001",
      "description": "Alternative Local Server (Port 5001)"
    }
  ],
...
  "paths": {
...
    "/api/device/{deviceId}/k100c": {
      "post": {
        "tags": ["IoT Device & Screens"],
        "summary": "Dispatch BLE Command to K100C Hardware Controller",
        "description": "Sends brightness, power switch, or display mode command via ESP32 BLE cloud bridge to the NovaStar/Colorlight/K100C LED sending card.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "power": { "type": "boolean", "example": true },
                  "brightness": { "type": "integer", "minimum": 0, "maximum": 100, "example": 85 },
                  "mode": { "type": "string", "example": "normal" },
                  "command": { "type": "string", "example": "SET_BRIGHTNESS:85" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Command queued and dispatched via WebSocket/BLE" }
        }
      }
    },
    "/api/device/{deviceId}/k100c/poll": {
      "get": {
        "tags": ["IoT Device & Screens"],
        "summary": "ESP32 WiFi Poll for K100C Commands",
        "description": "Onboard ESP32 microcontroller polls this endpoint periodically over WiFi to retrieve latest hardware state and BLE commands.",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Latest hardware state and commands",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deviceId": { "type": "string" },
                    "power": { "type": "boolean" },
                    "brightness": { "type": "integer" },
                    "mode": { "type": "string" },
                    "command": { "type": "string" },
                    "serverTime": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/screens": {
      "get": {
        "tags": ["Screens & Billboards Inventory"],
        "summary": "List All LED Screens",
        "responses": {
          "200": {
            "description": "List of screen displays",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Screen" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Screens & Billboards Inventory"],
        "summary": "Create Single LED Screen",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScreenInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Screen created" }
        }
      }
    },
    "/api/screens/pole": {
      "post": {
        "tags": ["Screens & Billboards Inventory"],
        "summary": "Provision Smart Pole with Dual Screens (Side A & Side B)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["poleId", "corridorName", "city", "area", "lat", "lng"],
                "properties": {
                  "poleId": { "type": "string", "example": "ETHREE-POLE3" },
                  "corridorName": { "type": "string", "example": "MG Road" },
                  "city": { "type": "string", "example": "Vijayawada" },
                  "area": { "type": "string", "example": "Benz Circle" },
                  "lat": { "type": "number", "example": 16.5062 },
                  "lng": { "type": "number", "example": 80.6480 },
                  "cctvCameraId": { "type": "string", "example": "sparsh-pole3" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Pole and both screens created successfully" }
        }
      }
    },
    "/api/screens/{id}": {
      "put": {
        "tags": ["Screens & Billboards Inventory"],
        "summary": "Update Screen",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Screen updated" } }
      },
      "delete": {
        "tags": ["Screens & Billboards Inventory"],
        "summary": "Delete Screen",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Screen deleted" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "Screen": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "deviceId": { "type": "string" },
          "name": { "type": "string" },
          "location": { "type": "string" },
          "corridorName": { "type": "string" },
          "poleId": { "type": "string" },
          "side": { "type": "string" },
          "status": { "type": "string" },
          "lat": { "type": "number" },
          "lng": { "type": "number" },
          "soundEnabled": { "type": "boolean" },
          "volume": { "type": "integer" }
        }
      },
      "ScreenInput": {
        "type": "object",
        "required": ["deviceId", "name", "location", "lat", "lng"],
        "properties": {
          "deviceId": { "type": "string" },
          "name": { "type": "string" },
          "location": { "type": "string" },
          "corridorName": { "type": "string" },
          "lat": { "type": "number" },
          "lng": { "type": "number" }
        }
      },
...
    }
  }
}